requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
frozenlist==1.4.1
gotrue==2.9.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
multidict==6.1.0
packaging==24.1
passlib==1.7.4
postgrest==0.17.1
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
realtime==2.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==13.1
yarl==1.15.2
